writeSetting: Ask board if loopback mode is valid
bladerf2 supports different loopback modes than the bladerf1.
version 2 of the library will error out if we pass a loopback mode
struct that is not valid for the attached device.

If we can, we ask the board if the loopback mode will work, and
log an error if it will not.

Signed-off-by: Josh Farwell <josh.farwell@gmail.com>