Implemented a more accurate hasGainMode
Presence/absence of LUT for AGC makes it actually a bit tough to tell whether or
not setGainMode will work, or error out as unsupported. This commit changes it
to check by attempting to set the gain mode, and then return true or false based
on the actual behavior. This way, hasGainMode is truly an indicator of whether
or not setGainMode is supported or not.