Reimplemented listSampleRates and listBandwidths
Though the range API is much better, these are still useful as sort of
"recommended" values for populating lists etc. I don't see anything that states
that SoapySDR has deprecated these - so they should still be included.

Namely, CubicSDR really needs these to be provided, as it relies on the lists to
populate several dropdowns. It seems appropriate for the Soapy module to provide
the list so that it is device-specific.