added saveQuickTune and reuseQuickTune args to setFrequency
This allows to save the bladerf_quick_tune (with args["saveQuickTune"]="1") when setting a specific frequency, and then reusing that bladerf_quick_tune when going back to that frequency (with args["reuseQuickTune"]="1").

It's also possible to schedule the frequency retune when reusing a bladerf_quick_tune by setting args["timestamp"] to the string representation of the desired timestamp.